[BLDC] Only enable rotor offset compensation when hall interpolation is enabled
The commit enabling sector offset compensation accidentally made some poor unit tests wherein the raw hall angle was not passed through if interpolation was disabled.

This feature should only be active if interpolation is also active, otherwise it does not make sense to actually enable this feature.

Fix adds a param flag for this and also fixes the unit tests to ensure raw hall angle passthrough